cluster: implement old dead keeper removal
add a DeadKeeperRemovalInterval (default to 48 hours) to the cluster
spec and implement dead keeper removal for keepers not healthy for more
than the specified interval. The current keeper assigned to the master
db will never be removed also if unhealthy or this will break the
cluster leaving it without a master.